AI/requirements.txt
Flask
langchain-core
langchain-groq
sentence-transformers
pandas
faiss-cpu
numpy
python-dotenv
accelerate>=0.26.0
watchdog